FoG is now using the recursive interface to access its lower layers
FoG accesses its lower layers with the same interface as it provides to
its higher layers. The old ILowerLayer interface is no longer in use.
Since the lower layer implementation (and not FoG) may use it in order
to implement its remote access via JINI, it was moved to the Bus
implementation.

The commit includes some compiler errors. The following issues are still
under construction:
- Rerouting and Rerouting experiment
- Lower layer implementation (Bus plug-ins)