Initial import of lower layer supporting the recursive interface
The layer does not support the remote access to media objects. It uses
the same command name "create bus" for creating the media objects. Thus,
it can not be used in parallel to the fog.bus plug-in (For debugging
purposes, the fog.bus project has to be closed in an Eclipse workspace
containing both projects).